Add files via upload
This file trains a neural network to distinguish between ten different types of beer (American IPA, Stout, etc). The characteristics used to classify are only:
   1) Alcohol by volume
   2) International bitterness units
   3) Volume of can

With this, I was able to achieve a test accuracy of about 55% to 62%. I believe, given the data set, that this is a reasonable outcome. More importantly, it has given me insight into developing a working neural network.